refactor: move CharacterCard to Widgets/character_card.dart
